Added authentication

+++ backend/routes/posts.js
@@ -34,7 +34,8 @@ router.post("", checkAuth, multer({storage: storage}).single("image"),
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(result => {
     res.status(201).json({
@@ -60,12 +61,18 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
   const post = new Post({
     _id: req.body.id,
     title: req.body.title,
-    content: req.body.content
+    content: req.body.content,
+    imagePath: imagePath,
+    creator: req.userData.userId
   })
-  Post.updateOne({_id: req.params.id}, post)
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post)
     .then(result => {
-      console.log(result);
-      res.status(200).json({message: 'Update successful!'});
+      if (result.nModified > 0){
+        res.status(200).json({message: 'Update successful!'});
+      } else {
+        res.status(401).json({message: 'Not authorized!'});
+      }
+
     });
 })
 
@@ -108,9 +115,13 @@ router.get(":id", (req, res, next)=>{
 router.delete("/:id", checkAuth, (req, res, next) => {
   // console.log(req.params.id);
 
-  Post.deleteOne({_id: req.params.id}).then(result => {
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
     console.log(result);
-    res.status(200).json( {message: "Post deleted!" });
+    if (result.n > 0){
+      res.status(200).json({message: 'Delete successful!'});
+    } else {
+      res.status(401).json({message: 'Not authorized!'});
+    }
   });
 
 });
